+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- `convertCSVtoJSON` now returns a JSON object
+
 ### Deprecated
 
 ### Removed